Added Analogous, Triadic, Tetradic Features
I made a hex conversion function to convert r g b into hex code for the computer to know which colors to display on the labels. Each of the function for the different categories are based off of different formulas to determine the hue angles, then using that angle I would convert it to hex code.